Convert token OG image template to JSX
- Exempt opengraph-image templates from the tsx boundary check in buildServer
- Omit the caption entirely when a token has no issuer
