+++ packages/frontend/src/components/opengraph-image/Token.tsx
@@ -1,0 +1,119 @@
+interface Props {
+  backgroundSrc: string
+  iconSrc: string
+  symbol: string
+  caption: string | undefined
+  size: {
+    width: number
+    height: number
+  }
+}
+
+// Unlike the other templates in this directory, this one is rendered by the
+// express server at runtime (see getInteropTokenOgImage), which is why it is
+// exempted from the tsx boundary check in scripts/buildServer.ts.
+export function TokenOpengraphImage({
+  backgroundSrc,
+  iconSrc,
+  symbol,
+  caption,
+  size,
+}: Props) {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
+    >
+      <img src={backgroundSrc} alt="" {...size} />
+
+      <div
+        style={{
+          position: 'absolute',
+          top: 100,
+          left: '50%',
+          transform: 'translateX(-50%)',
+          display: 'flex',
+          flexDirection: 'column',
+          justifyContent: 'center',
+          alignItems: 'center',
+          gap: 24,
+        }}
+      >
+        <svg
+          width="176"
+          height="72"
+          viewBox="0 0 176 72"
+          fill="none"
+          style={{
+            marginBottom: 32,
+          }}
+        >
+          <path
+            d="M31.366 3.051c3.226 1.893 6.074 4.323 7.667 7.77l13.062-7.77 15.233 7.312 3.465 13.367L39.61 62.598l-6.65 6.335-24.851-30.46a46 46 0 01-3.744-4.941 26.7 26.7 0 01-3.086-5.738C.422 25.782.004 23.81.004 21.858c0-4.443.836-8.268 2.549-11.495 1.633-3.228 3.982-5.718 7.01-7.49 3.046-1.754 6.63-2.65 10.772-2.65 4.082 0 7.746.956 11.031 2.828"
+            fill="#BD114F"
+          />
+          <path d="M34.155 34.588h-20.67v25.62h20.67z" fill="#BD114F" />
+          <path
+            d="M33.776 70.128c.916.897 2.05 1.355 3.385 1.355l33.751-.478c1.355 0 2.59-.18 3.386-1.375 1.035-1.553-1.713-7.51-5.337-7.928-3.604-.439-13.74-.459-19.514-.977-.458-.04-.578-.697-.418-1.135.418-1.235 17.224-21.137 17.224-21.137 1.274-1.455 2.529-3.108 3.743-4.921a28.3 28.3 0 003.107-5.757c.836-2.013 1.274-3.985 1.274-5.937 0-4.443-.856-8.268-2.568-11.495-1.633-3.227-3.983-5.718-7.01-7.47-3.046-1.774-6.63-2.65-10.772-2.65-4.062 0-7.746.936-11.032 2.828-3.226 1.873-6.073 4.323-7.666 7.75-.936 2.012-1.474 5.28-.936 7.49.458 1.814.916 2.272 2.29 2.73s1.812-.538 2.27-2.73c.458-2.19 1.214-4.661 2.29-6.394 1.374-2.192 3.903-3.805 5.934-4.562 1.712-.638 4.56-1.176 6.85-.917 4.082.458 6.97 1.913 9.1 4.383 2.15 2.47 3.803 6.574 3.325 9.682l-.02.1c-.279 1.912-.677 4.522-1.633 6.375-.975 1.892-3.942 5.618-5.774 7.75l-25.25 28.846c-.836.857-1.274 1.933-1.274 3.287 0 1.275.438 2.371 1.275 3.287"
+            fill="#FAFAFA"
+          />
+          <path
+            d="M8.766 45.226v-2.291c.06-1.375.08-2.909.08-4.602.06-1.753.06-3.466 0-5.1 0-1.693-.06-3.008-.16-3.944a20 20 0 01-.08-3.208c.1-.916.16-1.972.16-3.108 0-.777.02-1.454.08-2.052.06-.657.199-1.215.418-1.653a1.96 1.96 0 011.055-1.056c.498-.279 1.215-.418 2.131-.418.876 0 1.732.299 2.549.896.876.618 1.394 1.435 1.553 2.47l1.235 16.336.258 20.878 10.992-.338c.378 0 .936.2 1.653.578.756.318 1.135.737 1.135 1.235 0 .438-.18.817-.578 1.135a6 6 0 01-1.393.757c-.478.16-1.016.3-1.554.399-.498.06-.856.08-1.075.08q-.657.09-1.971.179h-7.467c-.538.259-1.195.378-1.972.319-.756-.06-1.394-.08-1.872-.08h-2.548c-.438 0-.936-.12-1.474-.339-.498-.259-.876-.558-1.155-.896z"
+            fill="#FAFAFA"
+          />
+          <path
+            d="M80.172 58.194c0-1.036.02-2.033.08-2.969.1-.916.159-1.872.159-2.869v-2.47c0-.219-.02-.598-.08-1.135a56 56 0 01-.16-1.813c0-.658-.039-1.255-.099-1.813-.04-.538-.08-.917-.08-1.136V27.056c0-1.534.08-3.009.26-4.423.219-1.474.258-2.928.159-4.363 1.095-1.255 2.41-2.012 3.942-2.29a21.6 21.6 0 014.68-.499c2.29 0 4.38.458 6.233 1.395 1.871.876 3.425 2.47 4.679 4.76.936 1.814 1.593 3.507 1.971 5.1.379 1.535.578 3.308.578 5.34v2.291a32 32 0 01-1.474 3.785c-.497 1.096-1.294 2.112-2.39 3.028a9.5 9.5 0 01-1.393.837c-.538.22-.936.458-1.155.737 2.09 1.195 3.584 2.63 4.52 4.263.996 1.654 1.474 3.626 1.474 5.917v2.132c-.1.657-.379 1.394-.817 2.231q-.657 1.136-1.314 1.793c-1.095.936-2.05 1.614-2.867 2.052s-1.653.777-2.47.996c-.816.22-1.692.319-2.628.319h-6.312c-.438-.1-.777-.22-.996-.319-.16-.12-.319-.219-.478-.339a1.26 1.26 0 00-.498-.398c-.159-.18-.438-.339-.816-.498-.338-.16-.637-.279-.916-.339-.219-.04-.438-.1-.657-.16-.16-.059-.359-.139-.578-.238a3.1 3.1 0 01-.557-.578zm6.89-1.315.159 1.056c.338.12.637.259.916.418.259.12.518.22.737.339q.329.15.657.239h.16c.935 0 1.85-.359 2.787-1.076.936-.697 1.394-1.634 1.394-2.79 0-.597.06-1.234.159-1.892.18-.657.12-1.633-.16-2.948-.278-1.594-1.035-2.77-2.29-3.526-1.274-.777-2.548-1.156-3.862-1.156h-.498a70 70 0 00-.16 4.264v7.072m0-16.017c2.35-.937 4.002-2.212 4.918-3.865.996-1.634 1.474-3.686 1.474-6.156 0-.937-.02-2.072-.08-3.447a18 18 0 00-.558-3.944c-.338-1.255-.856-2.331-1.573-3.208-.717-.936-1.712-1.394-3.026-1.394q-1.733 0-3.465.657v.658c0 .876.1 1.693.259 2.47.219.757.378 1.474.498 2.132zm19.355 13.885c-.12-2.31-.319-4.721-.578-7.231-.219-2.53-.498-4.921-.816-7.232V27.056c.04-.538.079-1.136.079-1.793.04-.658.1-1.295.16-1.893.059-.657.099-1.116.159-1.394.159-1.375.856-2.391 2.051-3.048 1.254-.658 2.688-1.096 4.281-1.315a25.4 25.4 0 014.919-.319h3.942-.498q.569 0 1.554.239 1.075.18 1.075.896c0 .399-.159.698-.498.917-.259.16-.597.239-.976.239 0 0-.338.06-.995.16-.657.06-1.414.179-2.29.338-.817.1-1.653.24-2.469.418-.757.1-1.235.18-1.394.24-.717.159-1.175.537-1.394 1.155v4.263c.06.877.119 1.992.159 3.367.12 1.375.199 2.73.259 4.104.04 1.315.1 2.49.159 3.526.06.996.08 1.514.08 1.574l.339.896c1.075 0 2.21-.02 3.365-.08 1.135-.099 2.21-.199 3.186-.238.836-.06 1.573.08 2.23.398.697.339 1.055.777 1.055 1.315 0 .438-.159.797-.498 1.076-.258.219-.617.398-1.055.577-.438.1-.876.22-1.314.32-.438.059-.777.119-.996.159-.816.12-1.414.219-1.792.338-.378.06-.737.1-1.075.16h-2.947c0 2.131-.04 4.243-.08 6.335-.06 2.012-.079 4.104-.079 6.235h1.871c.438-.06.857-.08 1.235-.08.378-.06.796-.08 1.234-.08.498-.059 1.155-.079 1.972-.079.279 0 .657.02 1.155.08a6.7 6.7 0 011.553.319c.538.12 1.015.298 1.394.577.378.28.577.658.577 1.156 0 .438-.159.817-.498 1.135a4.8 4.8 0 01-1.234.758l-1.314.478c-.438.12-.757.199-.976.259h-7.726a6.3 6.3 0 00-1.135-.08c-.398-.06-.777-.1-1.155-.1-.339-.04-.578-.08-.737-.08-.597 0-1.254-.099-1.971-.318-.717-.28-1.235-.658-1.553-1.156zm19.175 5.16c0-.916.06-2.172.16-3.765.179-1.654.338-3.228.498-4.762.159-1.593.298-2.968.418-4.104.159-1.155.239-1.753.239-1.812l3.126-23.09c0-.04.02-.259.08-.657.099-.379.219-.817.318-1.315.12-.478.219-.957.339-1.395.099-.478.199-.837.239-1.056.059-.279.119-.577.179-.916.099-.378.199-.737.318-1.056.1-.338.279-.597.498-.817s.518-.338.896-.338h3.127q.657 0 1.135.657c.338.438.577.857.756 1.235.16.498.379 1.116.657 1.893.319.697.657 1.414.976 2.131.339.718.597 1.375.816 1.973.22.538.339.836.339.896a3383 3383 0 005.576 18.07 181 181 0 014.52 17.331v.419c0 .876-.418 1.534-1.235 1.972-.757.438-1.454.657-2.051.657h-1.075c-.478-.06-1.016-.12-1.553-.179a5.5 5.5 0 01-1.394-.478c-.379-.22-.598-.518-.657-.916l-2.31-12.65c-1.075 0-2.43.04-4.023.099-1.513.04-2.867.08-4.022.08l-.318.239a97 97 0 01-1.554 13.068l-.338.478c-.1.12-.359.359-.737.737a10 10 0 01-.657.658h-.239c-.876 0-1.613-.24-2.23-.737-.538-.478-.817-1.036-.817-1.634zm13.72-16.575v-1.235c-.06-.22-.139-.578-.259-1.076-.04-.538-.139-1.096-.239-1.634-.119-.557-.219-1.036-.338-1.474-.04-.498-.1-.837-.159-.996-.1-.498-.319-1.255-.658-2.291a48 48 0 00-.975-3.367c-.339-1.155-.657-2.251-.996-3.287-.319-1.056-.577-1.813-.737-2.31l-2.29 17.67zm19.156 15.599v-5.18c.06-1.474.1-3.108.1-4.92.039-1.814.099-3.706.159-5.678.06-2.012.099-3.925.159-5.738s.12-3.426.159-4.84c.06-1.435.12-2.491.18-3.208 0-1.375.079-2.71.239-4.025a20.5 20.5 0 00.239-4.183h-6.651c-.379 0-.817-.1-1.314-.26-.439-.218-.658-.537-.658-.975 0-.399.1-.658.259-.837.219-.16.538-.319.976-.478 1.653 0 3.206-.06 4.679-.18 1.534-.1 3.047-.219 4.521-.318 1.473-.16 2.986-.299 4.52-.419 1.533-.1 3.146-.159 4.839-.159h2.787l1.573.817v.578c0 .557-.338.936-.995 1.155-.598.22-1.334.379-2.211.498-1.095.16-2.25.22-3.444.16-1.155-.06-1.932 0-2.31.159v6.893c0 3.128.079 6.215.259 9.283.159 3.068.239 6.077.239 9.045v4.024a79 79 0 01.079 3.686c0 1.215-.06 2.41-.159 3.625a36 36 0 01-.418 3.865 3.4 3.4 0 01-1.056.817c-.438.16-.856.299-1.234.398h-3.943c-.657-.04-1.175-.458-1.573-1.215z"
+            fill="#FAFAFA"
+          />
+        </svg>
+
+        <div
+          style={{
+            display: 'flex',
+            flexDirection: 'row',
+            justifyContent: 'center',
+            alignItems: 'center',
+            gap: 32,
+          }}
+        >
+          <img
+            src={iconSrc}
+            alt={`${symbol} logo`}
+            width={104}
+            height={104}
+            style={{ borderRadius: '50%' }}
+          />
+          <div
+            style={{
+              color: 'white',
+              whiteSpace: 'pre',
+              fontFamily: 'roboto',
+              fontWeight: 700,
+              fontSize: symbol.length < 22 ? 96 : 72,
+              lineHeight: 1,
+            }}
+          >
+            {symbol}
+          </div>
+        </div>
+        {caption && (
+          <div
+            style={{
+              color: '#db8bf7',
+              fontFamily: 'roboto',
+              fontWeight: 500,
+              fontSize: 36,
+              whiteSpace: 'pre',
+            }}
+          >
+            {caption}
+          </div>
+        )}
+      </div>
+    </div>
+  )
+}
